+++ src/test/java/eu/unicore/uftp/standalone/DispatcherTest.java
@@ -12,7 +12,7 @@ import eu.unicore.uftp.standalone.commands.ICommand;
 public class DispatcherTest {
 
 	@Test
-	public void testGetCommands() {
+	public void testGetCommands() throws Exception {
 		Collection<ICommand>commands = new ClientDispatcher().getCommands();
 		assertNotNull(commands);
 		assertTrue(commands.size()>0);