+++ src/test/java/eu/unicore/uftp/standalone/ssh/TestAgent.java
@@ -1,14 +1,14 @@
 package eu.unicore.uftp.standalone.ssh;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import java.io.File;
+import java.security.PublicKey;
+import java.util.ArrayList;
+import java.util.Collection;
 
-import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-import eu.unicore.services.restclient.sshkey.SSHKeyUC;
+import eu.unicore.services.restclient.jwt.JWTUtils;
 import eu.unicore.services.restclient.sshkey.SSHUtils;
 
 /**
@@ -18,28 +18,32 @@ public class TestAgent {
 	
 	@Test
 	@Disabled
-	public void testSigningUsingAgent() throws Exception {
-		String token = "test123";
+	public void testSigningUsingAgentRSA() throws Exception {
 		String user = "demouser";
-		File privKey = new File(System.getProperty("user.home")+"/.ssh/id_rsa");
-		SshKeyHandler handler = new SshKeyHandler(privKey, user, token, true);
+		File privKey = new File(System.getProperty("user.home")+"/.ssh/id_rsa_new");
+		SshKeyHandler handler = new SshKeyHandler(privKey, user, true);
 		handler.selectIdentity();
-		SSHKeyUC authData = (SSHKeyUC)handler.getAuthData();
-		String pubKey = FileUtils.readFileToString(new File(System.getProperty("user.home")+"/.ssh/id_rsa.pub"), "UTF-8");
-		assertTrue(SSHUtils.validateAuthData(authData, pubKey));
+		SSHAgentKey authData = (SSHAgentKey)handler.getAuthData();
+		String token = authData.getToken();
+		Collection<PublicKey> pubKeys = new ArrayList<>();
+		pubKeys.add(SSHUtils.readPublicKey(new File(System.getProperty("user.home")+
+				"/.ssh/id_rsa_new.pub")));
+		JWTUtils.verifyJWTToken(token, pubKeys, null);
 	}
 
 	@Test
 	@Disabled
 	public void testSigningUsingAgentEd25519() throws Exception {
-		String token = "test123";
 		String user = "demouser";
 		File privKey = new File(System.getProperty("user.home")+"/.ssh/id_ed25519");
-		SshKeyHandler handler = new SshKeyHandler(privKey, user, token, true);
+		SshKeyHandler handler = new SshKeyHandler(privKey, user, true);
 		handler.selectIdentity();
-		SSHKeyUC authData = (SSHKeyUC)handler.getAuthData();
-		String pubKey = FileUtils.readFileToString(new File(System.getProperty("user.home")+"/.ssh/id_ed25519.pub"), "UTF-8");
-		assertTrue(SSHUtils.validateAuthData(authData, pubKey));
+		SSHAgentKey authData = (SSHAgentKey)handler.getAuthData();
+		String token = authData.getToken();
+		Collection<PublicKey> pubKeys = new ArrayList<>();
+		pubKeys.add(SSHUtils.readPublicKey(new File(System.getProperty("user.home")+
+				"/.ssh/id_ed25519.pub")));
+		JWTUtils.verifyJWTToken(token, pubKeys, null);
 	}
 
 }